+++ src/events.h
@@ -3,6 +3,7 @@
 enum CustomEvent {
   about_dialog_requested,
   change_unit_system,
+  display_alert,
   model_updated,
   new_location_requested,
   refresh_requested,