+++ src/util.h
@@ -2,7 +2,10 @@
 
 #include <algorithm>
 #include <cctype>
+#include <chrono>
+#include <ctime>
 #include <locale>
+#include <memory>
 #include <sstream>
 #include <stdexcept>
 #include <string>
@@ -10,6 +13,8 @@
 
 #include "errors.h"
 #include "model.h"
+
+using namespace std::chrono_literals;
 
 namespace taranis {
 std::vector<double>
@@ -50,5 +55,34 @@ inline void rtrim(std::string &s) {
                    [](char ch) { return not std::isspace(ch) and ch != ','; })
           .base(),
       s.end());
+}
+
+constexpr char time_format[] = "%H:%M";
+constexpr char full_date_format[] = "%R (%F)";
+static char formatted_time[100];
+
+inline std::string format_time(const std::time_t &time, bool full = false) {
+  auto format = full ? full_date_format : time_format;
+  std::strftime(formatted_time, sizeof(formatted_time), format,
+                std::localtime(&time));
+  // TODO should use GetLangTime() to use user "locale" but don't know
+  // how it works…
+  return formatted_time;
+}
+
+inline std::string format_duration(const std::time_t &start,
+                                   const std::time_t &end) {
+  const auto duration =
+      std::chrono::duration_cast<std::chrono::hours>((end - start) * 1s);
+  if (duration < 1h) {
+    return "<1h";
+  }
+  std::stringstream duration_text;
+  duration_text << static_cast<int>(
+                       std::chrono::duration_cast<std::chrono::hours>(
+                           (end - start) * 1s)
+                           .count())
+                << "h";
+  return duration_text.str();
 }
 } // namespace taranis